@@ -85,8 +85,8 @@
     "pino-pretty": "7.6.0",
     "prop-types": "15.8.1",
     "proxy-agent": "5.0.0",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
+    "react": "18.0.0",
+    "react-dom": "18.0.0",
     "react-flow-renderer": "10.0.7",
     "react-icons": "4.3.1",
     "react-json-view": "1.21.3",
